requirements.txt
asgiref==3.6.0
distro-info==1.0
Django==4.1.7
django-cors-headers==4.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==20.1.0
inflection==0.5.1
packaging==23.1
psycopg2-binary==2.9.7
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.3
uritemplate==4.1.1
